# Improve `Message` sorting performance