@MichaReiser 2023-05-24
Current dependencies on/for this PR:
* main
  * **PR #4624** <a href="https://app.graphite.dev/github/pr/charliermarsh/ruff/4624" target="_blank"><img src="https://static.graphite.dev/graphite-32x32.png" alt="Graphite" width="10px" height="10px"/></a>  👈

This comment was auto-generated by [Graphite](https://app.graphite.dev/github/pr/charliermarsh/ruff/4624?utm_source=stack-comment).